requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
sqlalchemy>=2.0.41
pydantic-settings==2.1.0
python-multipart==0.0.21
python-dotenv==1.2.1
alembic==1.13.1
psycopg2-binary>=2.9.9